# documentation-swagger/swagger.yaml
openapi: 3.0.1
info:
  version: '1.0.0'
  title: REST API Documentation
  description: REST API Documentation
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /gift-certificates:
    get:
      tags:
        - gift-certificate-controller
      description: Returns all gift certificates
      summary: Returns all gift certificates
      operationId: getAllCertificates
      parameters:
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
    post:
      tags:
        - gift-certificate-controller
      description: Returns created gift certificate
      summary: Returns created gift certificate
      security:
        - BearerAdminAuth: []
      operationId: createCertificate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftCertificateForPostAndUpdate'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /gift-certificates/{id}:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificate by id
      summary: Returns gift certificate by id
      operationId: getGiftCertificateById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
    delete:
      tags:
        - gift-certificate-controller
      description: Returns status for deleting gift certificate
      summary: Returns status for deleting gift certificate
      security:
        - BearerAdminAuth: []
      operationId: deleteGiftCertificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  HTTP Status:
                    type: string
                    enum:
                      - OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
    patch:
      tags:
        - gift-certificate-controller
      description: Returns updated gift certificate
      summary: Returns updated gift certificate
      security:
        - BearerAdminAuth: []
      operationId: updateGiftCertificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GiftCertificateForPostAndUpdate'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /gift-certificates/update-price/{id}:
    patch:
      tags:
        - gift-certificate-controller
      description: Returns gift certificate with updated price
      summary: Returns gift certificate with updated price
      security:
        - BearerAdminAuth: []
      operationId: updatePrice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              minimum: 1
              type: integer
              format: int32
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /gift-certificates/update-duration/{id}:
    patch:
      tags:
        - gift-certificate-controller
      description: Returns gift certificate with updated duration
      summary: Returns gift certificate with updated duration
      security:
        - BearerAdminAuth: []
      operationId: updateDuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              minimum: 1
              type: integer
              format: int32
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificate'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus401'
          description: UNAUTHORIZED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus403'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus404'
          description: Item Not Found
  /gift-certificates/get-tag-name:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificates by tag name
      summary: Returns gift certificates by tag name
      operationId: getGiftCertificatesByTagName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
  /gift-certificates/get-sort-name:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificates sorted by name
      summary: Returns gift certificates sorted by name
      operationId: getGiftCertificatesSortedByName
      parameters:
        - name: sortingType
          in: query
          required: false
          schema:
            type: string
            default: ASC
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
  /gift-certificates/get-sort-name-date:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificates sorted by name and by date
      summary: Returns gift certificates sorted by name and by date
      operationId: getGiftCertificatesSortedByNameByDate
      parameters:
        - name: sortingTypeName
          in: query
          required: false
          schema:
            type: string
            default: ASC
        - name: sortingTypeDate
          in: query
          required: false
          schema:
            type: string
            default: ASC
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
  /gift-certificates/get-gift-certificate-description:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificates by part of description
      summary: Returns gift certificates by part of description
      operationId: getGiftCertificatesByPartOfDescription
      parameters:
        - name: description
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
  /gift-certificates/get-by-tag-names:
    get:
      tags:
        - gift-certificate-controller
      description: Returns gift certificates by tag names
      summary: Returns gift certificates by tag names
      operationId: getGiftCertificatesByTags
      parameters:
        - name: tagNamesSet
          in: query
          required: true
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
        - name: page
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftCertificates'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpStatus400'
          description: Bad Request
components:
  schemas:
...
    GiftCertificateForPostAndUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          minimum: 1
          type: integer
          format: int32
        duration:
          minimum: 1
          type: integer
          format: int32
        createDate:
          type: string
          format: date-time
        lastUpdateDate:
          type: string
          format: date-time
        tags:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/TagForPost'
...
    GiftCertificate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          minimum: 1
          type: integer
          format: int32
        duration:
          minimum: 1
          type: integer
          format: int32
        createDate:
          type: string
          format: date-time
        lastUpdateDate:
          type: string
          format: date-time
        tags:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        _links:
          $ref: '#/components/schemas/Links'
...
    GiftCertificates:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/GiftCertificate'
        _links:
          $ref: '#/components/schemas/Links'
...
    HttpStatus400:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - BAD_REQUEST
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "BAD_REQUEST"
    HttpStatus401:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - Unauthorized
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    HttpStatus403:
      type: object
      properties:
        HTTP Status:
          type: string
          enum:
            - FORBIDDEN
        response body:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "FORBIDDEN"
...
    Links:
      type: array
      items:
        type: object
        properties:
          rel:
            type: string
          href:
            type: string
...
  securitySchemes:

    BearerAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'For ADMIN'

    BearerUserOrAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'For USER or ADMIN'

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: See your primary Google Account email address
            profile: See your personal info, including any personal info you've made publicly available
            openid: Associate you with your personal info on Google